(chore) made seperate handlers for socket
